Refactor out UI
